+++ components/Profile/ProfileDetails.tsx
@@ -22,7 +22,7 @@ import LoginPage from "../../pages/login";
 import HeaderNav from "../HeaderNav/HeaderNav";
 import { UserData } from "../../utils/interfaces";
 import { GetCompany } from "./GetCompany";
-import useUser from "../TwaliContext";
+import useUser from "../../context/TwaliContext";
 
 const ProfileDetails = ({ user }) => {
   // Fallback for getStaticPaths, when fallback: true
