+++ components/reusable/TwaliSlider.tsx
@@ -8,7 +8,31 @@ import {
 } from "@chakra-ui/react";
 import React from "react";
 
-export function Current({ setCurrentStatus, currentStatus, defaultValue }) {
+interface Props {
+  setCurrentStatus: React.Dispatch<React.SetStateAction<number>>;
+  currentStatus: number;
+  defaultValue?: number;
+  marks: [string, string];
+}
+
+/**
+ * Twali slider component for binary selection
+ *
+ *
+ * @param setCurrentStatus  React.Dispatch<React.SetStateAction<number>>
+ * @param currentStatus tracks current status of slider. <Number> 0 || 1
+ * @param marks string array of length 2 containing the desired slider marks. string[]
+ * @param defaultValue optional <Number> 0 || 1. defaults to 0 if no value is provided
+ *
+ * @returns JSX slider element
+ *
+ */
+export function TwaliSlider({
+  setCurrentStatus,
+  currentStatus,
+  marks,
+  defaultValue,
+}: Props) {
   return (
     <Slider
       ml={2}
@@ -21,10 +45,10 @@ export function Current({ setCurrentStatus, currentStatus, defaultValue }) {
       onChange={(v) => setCurrentStatus(v)}
     >
       <SliderMark value={0} m="16px" ml="-2.5" fontSize="sm">
-        No
+        {marks[0]}
       </SliderMark>
       <SliderMark value={1.25} m="16px" ml="-2.5" fontSize="sm">
-        Yes
+        {marks[1]}
       </SliderMark>
       <SliderTrack bg="n2" height={"24px"} width={"24px"} borderRadius={"full"}>
         <Box position="relative" right={200} />
