+++ utils/setEventArray.tsx
